refactor: issues by reviewers
