+++ src/components/directives/bind-with/test/unit/main.ts
@@ -6,26 +6,18 @@
  * https://github.com/V4Fire/Client/blob/master/LICENSE
  */
 
+import type { JSHandle } from 'playwright';
+
 import test from 'tests/config/unit/test';
 import { Component } from 'tests/helpers';
-import type { JSHandle, Locator, Page } from 'playwright';
+
 import type iBlock from 'components/super/i-block/i-block';
-import type { Listener } from 'components/directives/bind-with';
+import {
 
-/**
- * A call to v-bind-with's .then() or .catch()
- */
-interface BindWithTestCallInfo {
-	args: any[];
-}
+	createDivForBindWithTest,
+	getBindWithTestInfo
 
-/**
- * A history of calls to v-bind-with's .then()/.catch()
- */
-interface BindWithTestInfo {
-	calls: BindWithTestCallInfo[];
-	errorCalls: BindWithTestCallInfo[];
-}
+} from 'components/directives/bind-with/test/helpers';
 
 test.describe('<div v-bind-with>', () => {
 
@@ -36,11 +28,7 @@ test.describe('<div v-bind-with>', () => {
 		rootHandle = (await Component.getComponentByQuery(page, '#root-component'))!;
 	});
 
-	test.afterEach(async ({page}) => {
-		await Component.removeCreatedComponents(page);
-	});
-
-	test('handler execution on event emission', async ({page}) => {
+	test('when event is emitted, handler is executed', async ({page}) => {
 		const divLocator = await createDivForBindWithTest(page, {
 			on: 'testEvent'
 		});
@@ -53,7 +41,7 @@ test.describe('<div v-bind-with>', () => {
 		test.expect(info!.calls.length).toBe(1);
 	});
 
-	test('handler execution on field change', async ({page}) => {
+	test('when field is changed, handler is executed', async ({page}) => {
 		const divLocator = await createDivForBindWithTest(page, {
 			path: 'testField'
 		});
@@ -66,7 +54,7 @@ test.describe('<div v-bind-with>', () => {
 		test.expect(info!.calls.length).toBe(1);
 	});
 
-	test('handler execution inside callback', async ({page}) => {
+	test('handler is executed as a callback', async ({page}) => {
 		const divLocator = await createDivForBindWithTest(page, {
 			callback: (handler) => [1].forEach(handler)
 		});
@@ -76,7 +64,7 @@ test.describe('<div v-bind-with>', () => {
 		test.expect(info!.calls[0].args).toStrictEqual([1, 0, [1]]);
 	});
 
-	test('handler execution on external emitter event', async ({page}) => {
+	test('when event is emitted by external emitter, handler is executed', async ({page}) => {
 		const divLocator = await createDivForBindWithTest(page, {
 			emitter: (event: string, listener: AnyFunction) => {
 				document.body.addEventListener(event, listener);
@@ -93,7 +81,7 @@ test.describe('<div v-bind-with>', () => {
 		test.expect(info!.calls.length).toBe(1);
 	});
 
-	test('handler execution on promise resolution', async ({page}) => {
+	test('when promise is resolved, handler is executed', async ({page}) => {
 		const divLocator = await createDivForBindWithTest(page, {
 			promise: () => Promise.resolve()
 		});
@@ -103,7 +91,7 @@ test.describe('<div v-bind-with>', () => {
 		test.expect(info!.errorCalls.length).toBe(0);
 	});
 
-	test('error handler execution on promise rejection', async ({page}) => {
+	test('when promise is rejected, error handler is executed', async ({page}) => {
 		const divLocator = await createDivForBindWithTest(page, {
 			promise: () => Promise.reject(new Error('rejection'))
 		});
@@ -113,7 +101,7 @@ test.describe('<div v-bind-with>', () => {
 		test.expect(info!.errorCalls.length).toBe(1);
 	});
 
-	test('single handler execution with `once` option', async ({page}) => {
+	test('when `once` option is set, handler is executed only once', async ({page}) => {
 		const divLocator = await createDivForBindWithTest(page, {
 			once: 'testEvent'
 		});
@@ -127,7 +115,7 @@ test.describe('<div v-bind-with>', () => {
 		test.expect(info!.calls.length).toBe(1);
 	});
 
-	test('correctness of handler arguments', async ({page}) => {
+	test('handler receives correct arguments', async ({page}) => {
 		const divLocator = await createDivForBindWithTest(page, {
 			on: 'onTestEvent'
 		});
@@ -140,7 +128,7 @@ test.describe('<div v-bind-with>', () => {
 		test.expect(info!.calls[0].args).toStrictEqual([1, 2, 3]);
 	});
 
-	test('multiple handler executions when array is passed', async ({page}) => {
+	test('when array is passed, handler is executed on every event', async ({page}) => {
 		const divLocator = await createDivForBindWithTest(page, [
 			{
 				once: 'testEvent'
@@ -158,85 +146,6 @@ test.describe('<div v-bind-with>', () => {
 		test.expect(info).toBeTruthy();
 		test.expect(info!.calls.length).toBe(2);
 	});
-
-	/**
-	 * A handler to pass as .then()/.catch() in v-bind-with
-	 *
-	 * @param element - The target element
-	 * @param args - Args provided by v-bind-with trigger (on/path/callback...)
-	 */
-	function handler(element: HTMLElement, ...args: any[]) {
-		const previousInfo: Partial<BindWithTestInfo> =
-			JSON.parse(element.getAttribute('data-test-bind-with') ?? '{}');
-		const newInfo: BindWithTestInfo = {
-			calls: previousInfo.calls ?? [],
-			errorCalls: previousInfo.errorCalls ?? []
-		};
-		const preparedArgs = args.map(
-			(arg: any) => {
-				// Avoid converting circular structure to JSON
-				try {
-					JSON.stringify(arg);
-				} catch (e) {
-					return null;
-				}
-
-				return arg;
-			}
-		);
-		let callDestination: keyof BindWithTestInfo = 'calls';
-		if (args.length > 0 && args[0] instanceof Error) {
-			callDestination = 'errorCalls';
-		}
-
-		newInfo[callDestination].push({
-			args: preparedArgs
-		});
-
-		element.setAttribute('data-test-bind-with', JSON.stringify(newInfo));
-	}
-
-	/**
-	 * Force put our handlers to given v-bind-with listener.
-	 * @param listener - A v-bind-with listener to process
-	 */
-	function addTestHandlersToListener(listener: Partial<Listener>) {
-		return {
-			...listener,
-			then: handler,
-			catch: handler
-		};
-	}
-
-	/**
-	 * Create a <div> with v-bind-with set by test code.
-	 *
-	 * @param page - The page.
-	 * @param bindWithValue - Value to pass to v-bind-with
-	 */
-	async function createDivForBindWithTest(page: Page, bindWithValue: CanArray<Partial<Listener>>) {
-		await Component.createComponent(page, 'div', {
-			'v-bind-with': Object.isArray(bindWithValue) ?
-				bindWithValue.map(addTestHandlersToListener) :
-				addTestHandlersToListener(bindWithValue),
-			'data-testid': 'div'
-		});
-
-		return page.getByTestId('div');
-	}
-
-	/**
-	 * Get v-bind-with calls info by given locator
-	 * @param locator - The source locator
-	 */
-	async function getBindWithTestInfo(locator: Locator): Promise<BindWithTestInfo | null> {
-		const attrValue = await locator.getAttribute('data-test-bind-with');
-		if (attrValue == null) {
-			return null;
-		}
-
-		return <BindWithTestInfo>JSON.parse(attrValue);
-	}
 
 });
