+++ src/components/pages/p-v4-components-demo/index.js
@@ -40,10 +40,10 @@ package('p-v4-components-demo')
 	)
 
 	.libs(
+		'components/directives/bind-with',
 		'components/directives/image',
 		'components/directives/icon',
 		'core/router/engines/browser-history',
 		'core/router/engines/in-memory',
 		'components/traits/i-control-list/i-control-list',
-		'components/directives/bind-with',
 	);
